prettier CLI fix, move constants to canvas
