+++ src/canvas/canvasUtils.js
@@ -13,7 +13,7 @@ import {
   REACT_WORK_SIZE,
   REACT_PRIORITY_BORDER_SIZE,
   FLAMECHART_FRAME_HEIGHT,
-} from '../constants';
+} from './constants';
 import usePanAndZoom, {
   durationToWidth,
   positionToTimestamp,
