+++ src/canvas/constants.js
